docs/pyproject.toml
# SPDX-FileCopyrightText: 2019-2020 Magenta ApS
# SPDX-License-Identifier: MPL-2.0
[tool.poetry]
name = "OS2mo docs"
version = "0.1.0"
description = "OS2mo - Medarbejder og Organisation"
authors = ["Magenta <info@magenta.dk>"]
license = "MPL-2.0"
documentation = "https://os2mo.readthedocs.io"
readme = "README.rst"
homepage = "https://magenta.dk/"
repository = "https://git.magenta.dk/rammearkitektur/os2mo"
keywords = ["os2mo", "docs", "documentation"]

[tool.poetry.dependencies]
python = "^3.8"
mkdocs-material = "^8.1.4"
graphql-core = "^3.1.7"
Jinja2 = "^3.0.3"
pydantic = "^1.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autopub]
git-username = "MagentaBot"
git-email = "info@magenta.dk"
